Re-architected Assumption class
Pivoted the assumption data in a way that made sense. Re-architected the dateTime filter in preparation for a new "re-filter" button. The intention is that the re-filter button will not have to re-query the whole scenario, just re-create the pandas df. Fixed a bug where an XOR of start/end would crash the system.